First draft of bandwidth tester Comitted together with binary
This is a first draft of an internet download speed meter.
It works by downloading a file of a fixed size from
https://www.thinkbroadband.com/download and then calculating how long it
took.

The download is executed in a loop. Beware if you are on a data plan as
this can consume a lot of data. Also the provider of these files,
thinkbroadband.com mentions blocking repetitive attempts.